Rename subscriptions manager and worker lambdas
